Update equals/hashcode implementations for aggregations
