+++ kaldb/src/main/java/com/slack/kaldb/logstore/search/aggregations/UniqueCountAggBuilder.java
@@ -24,37 +24,18 @@ public class UniqueCountAggBuilder extends ValueSourceAggBuilder {
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
+    if (!(o instanceof UniqueCountAggBuilder)) return false;
+    if (!super.equals(o)) return false;
 
     UniqueCountAggBuilder that = (UniqueCountAggBuilder) o;
 
-    // AggBuilderBase
-    if (!Objects.equals(super.name, that.name)) return false;
-    if (!Objects.equals(super.metadata, that.metadata)) return false;
-    if (!Objects.equals(super.subAggregations, that.subAggregations)) return false;
-
-    // ValueSourceAggBuilder
-    if (!Objects.equals(super.field, that.field)) return false;
-    if (!Objects.equals(super.missing, that.missing)) return false;
-
-    // UniqueCountAggBuilder
     return Objects.equals(precisionThreshold, that.precisionThreshold);
   }
 
   @Override
   public int hashCode() {
-    // UniqueCountAggBuilder
-    int result = precisionThreshold != null ? precisionThreshold.hashCode() : 0;
-
-    // ValueSourceAggBuilder
-    result = 31 * result + (field != null ? field.hashCode() : 0);
-    result = 31 * result + (missing != null ? missing.hashCode() : 0);
-
-    // AggBuilderBase
-    result = 31 * result + (name != null ? name.hashCode() : 0);
-    result = 31 * result + (metadata != null ? metadata.hashCode() : 0);
-    result = 31 * result + (subAggregations != null ? subAggregations.hashCode() : 0);
-
+    int result = super.hashCode();
+    result = 31 * result + (precisionThreshold != null ? precisionThreshold.hashCode() : 0);
     return result;
   }
